Persist cartId throughout the App
THis POC uses a custom hook in order to persist the cartId throughout
the application.
The entire logic is encapsulated within the hook and it will either draw
a value from localStorage (a new service that needs testing) or generate
one.

This hook can be reused throughout the app as once it passed the initial
id generation, it will draw it from localStorage.